Fix configuration to allow spin-up
Here, the main concern is having an actual runoff field, generated by:

  python3 regrid_runoff.py hgrid_01.nc ocean_mask.nc -m mask \
    RYF.runoff_all.1990_1991.nc -r friver eac_runoff.nc
  ncatted -O -a cartesian_axis,i,a,c,X -a cartesian_axis,j,a,c,Y \
    eac_runoff.nc

But, if the field wasn't present, the issue is also fixed by setting
the "factor" (last) field for runoff to 0, rather than 1 in the
data_table.

We've also included initial velocity/eta conditions, which aren't
strictly necessary for the spinup, but get to a balanced state much
more quickly.